Fixed CommandReturnObject::SetImmediateErrorFile() to set the correct stream.
Modifed lldb_private::Process to be able to handle connecting to a remote 
target that isn't running a process. This leaves lldb_private::Process in the
eStateConnected state from which we can then do an attach or launch.

Modified ProcessGDBRemote to be able to set stdin, stdout, stderr, working
dir, disable ASLR and a few other settings down by using new GDB remote 
packets. This allows us to keep all of our current launch flags and settings
intact and still be able to communicate them over to the remote GDB server.
Previously these were being sent as arguments to the debugserver binary that
we were spawning. Also modified ProcessGDBRemote to handle losing connection
to the remote GDB server and always exit immediately. We do this by watching
the lldb_private::Communication event bit for the read thread exiting in the
ProcessGDBRemote async thread.

Added support for many of the new 'Q' packets for setting stdin, stdout,
stderr, working dir and disable ASLR to the GDBRemoteCommunication class for
easy accesss.

Modified debugserver for all of the new 'Q' packets and also made it so that
debugserver always exists if it loses connection with the remote debugger.

llvm-svn: 126444

+++ lldb/source/Commands/CommandObjectProcess.cpp
@@ -163,51 +163,68 @@ public:
         // If our listener is NULL, users aren't allows to launch
         char filename[PATH_MAX];
         const Module *exe_module = target->GetExecutableModule().get();
+
+        if (exe_module == NULL)
+        {
+            result.AppendError ("no file in target, set executable file using 'file' command");
+            result.SetStatus (eReturnStatusFailed);
+            return false;
+        }
+        
         exe_module->GetFileSpec().GetPath(filename, sizeof(filename));
 
+        StateType state = eStateInvalid;
         Process *process = m_interpreter.GetDebugger().GetExecutionContext().process;
-        if (process && process->IsAlive())
-        {        
-            char message[1024];
-            if (process->GetState() == eStateAttaching)
-                ::strncpy (message, "There is a pending attach, abort it and launch a new process?", sizeof(message));
-            else
-                ::strncpy (message, "There is a running process, kill it and restart?", sizeof(message));
-    
-            if (!m_interpreter.Confirm (message, true))
-            {
-                result.SetStatus (eReturnStatusFailed);
-                return false;
-            }
-            else
-            {
-                Error error (process->Destroy());
-                if (error.Success())
+        if (process)
+        {
+            state = process->GetState();
+            
+            if (process->IsAlive() && state != eStateConnected)
+            {       
+                char message[1024];
+                if (process->GetState() == eStateAttaching)
+                    ::strncpy (message, "There is a pending attach, abort it and launch a new process?", sizeof(message));
+                else
+                    ::strncpy (message, "There is a running process, kill it and restart?", sizeof(message));
+        
+                if (!m_interpreter.Confirm (message, true))
                 {
-                    result.SetStatus (eReturnStatusSuccessFinishResult);
+                    result.SetStatus (eReturnStatusFailed);
+                    return false;
                 }
                 else
                 {
-                    result.AppendErrorWithFormat ("Failed to kill process: %s\n", error.AsCString());
-                    result.SetStatus (eReturnStatusFailed);
+                    Error error (process->Destroy());
+                    if (error.Success())
+                    {
+                        result.SetStatus (eReturnStatusSuccessFinishResult);
+                    }
+                    else
+                    {
+                        result.AppendErrorWithFormat ("Failed to kill process: %s\n", error.AsCString());
+                        result.SetStatus (eReturnStatusFailed);
+                    }
                 }
             }
         }
         
-        const char *plugin_name;
-        if (!m_options.plugin_name.empty())
-            plugin_name = m_options.plugin_name.c_str();
-        else
-            plugin_name = NULL;
-
-        process = target->CreateProcess (m_interpreter.GetDebugger().GetListener(), plugin_name).get();
-
-        if (process == NULL)
+        if (state != eStateConnected)
         {
-            result.AppendErrorWithFormat ("Failed to find a process plugin for executable.\n");
-            result.SetStatus (eReturnStatusFailed);
-            return false;
+            const char *plugin_name;
+            if (!m_options.plugin_name.empty())
+                plugin_name = m_options.plugin_name.c_str();
+            else
+                plugin_name = NULL;
+
+            process = target->CreateProcess (m_interpreter.GetDebugger().GetListener(), plugin_name).get();
+            if (process == NULL)
+            {
+                result.AppendErrorWithFormat ("Failed to find a process plugin for executable.\n");
+                result.SetStatus (eReturnStatusFailed);
+                return false;
+            }
         }
+
 
         // If no launch args were given on the command line, then use any that
         // might have been set using the "run-args" set variable.
@@ -219,16 +236,24 @@ public:
         
         if (m_options.in_new_tty)
         {
-            char exec_file_path[PATH_MAX];
-            if (exe_module->GetFileSpec().GetPath(exec_file_path, sizeof(exec_file_path)))
+            if (state == eStateConnected)
             {
-                launch_args.InsertArgumentAtIndex(0, exec_file_path);
+                result.AppendWarning("launch in tty option is ignored when launching through a remote connection");
+                m_options.in_new_tty = false;
             }
             else
             {
-                result.AppendError("invalid executable");
-                result.SetStatus (eReturnStatusFailed);
-                return false;
+                char exec_file_path[PATH_MAX];
+                if (exe_module->GetFileSpec().GetPath(exec_file_path, sizeof(exec_file_path)))
+                {
+                    launch_args.InsertArgumentAtIndex(0, exec_file_path);
+                }
+                else
+                {
+                    result.AppendError("invalid executable");
+                    result.SetStatus (eReturnStatusFailed);
+                    return false;
+                }
             }
         }
 
@@ -563,9 +588,11 @@ public:
         bool synchronous_execution = m_interpreter.GetSynchronous ();
         
         Process *process = m_interpreter.GetDebugger().GetExecutionContext().process;
+        StateType state = eStateInvalid;
         if (process)
         {
-            if (process->IsAlive())
+            state = process->GetState();
+            if (process->IsAlive() && state != eStateConnected)
             {
                 result.AppendErrorWithFormat ("Process %u is currently being debugged, kill the process before attaching.\n", 
                                               process->GetID());
@@ -610,12 +637,15 @@ public:
         }
         else
         {
-            const char *plugin_name = NULL;
-            
-            if (!m_options.plugin_name.empty())
-                plugin_name = m_options.plugin_name.c_str();
+            if (state != eStateConnected)
+            {
+                const char *plugin_name = NULL;
+                
+                if (!m_options.plugin_name.empty())
+                    plugin_name = m_options.plugin_name.c_str();
 
-            process = target->CreateProcess (m_interpreter.GetDebugger().GetListener(), plugin_name).get();
+                process = target->CreateProcess (m_interpreter.GetDebugger().GetListener(), plugin_name).get();
+            }
 
             if (process)
             {
@@ -1503,7 +1533,10 @@ public:
                 }
                 else
                 {
-                    output_stream.Printf ("Process %d %s\n", exe_ctx.process->GetID(), StateAsCString (state));
+                    if (state == eStateConnected)
+                        output_stream.Printf ("Connected to remote target.\n");
+                    else
+                        output_stream.Printf ("Process %d %s\n", exe_ctx.process->GetID(), StateAsCString (state));
                     if (exe_ctx.thread == NULL)
                         exe_ctx.thread = exe_ctx.process->GetThreadList().GetThreadAtIndex(0).get();
                     if (exe_ctx.thread != NULL)
